mcp/client/requirements.txt
fastapi
uvicorn[all]
loguru
mcp[cli]
python-dotenv
python-socketio
dotenv
websocket-client
requests